arista: Add spr.autoUpdateMessage default true
Add a config message to default to always updating the github
title + message from the commit.

This differs from the previous default, which warned the user
that the PR and the commit message are out-of-sync, requiring
the user to run manual steps to update in either direction.

Pull-Request: https://github.com/aristanetworks/cordspr/pull/5