arista: Make commit sections "nearly" trailer-safe (BREAKING CHANGE)
Change the commit section names to be case-sensitive, and not
contain whitespace.  This is more compatible with
git commit trailers, so they may interop to some extent.

Commit trailers are still somewhat brittle; the overall
commit message formatting has not been revisited to bring
it all in line with git's formatting for trailers.

Pull-Request: https://github.com/aristanetworks/cordspr/pull/7