arista: add new commit_message.rs file to parse commit messages
We use the 'git interpret-trailers --parse' command to properly
parse the commit trailers.

We'll use it to replace the message::parse_message() implementation,
which doesn't adhere to the official git commit trailers, as described
in:

   https://git-scm.com/docs/git-interpret-trailers

We'll change this implementation because it has several problems with
commit trailers (e.g. silently discards trailers that spr is not aware
of, etc.)

Pull-Request: https://github.com/aristanetworks/cordspr/pull/9