arista: modify parse_message() to be trailer-safe (BREAKING CHANGE)
Problem Definition
------------------

Before these changes, the parse_message() function used an ad hoc
logic to parse git commit messages that was not compliant with the
official trailer syntax, as described in:

  https://git-scm.com/docs/git-interpret-trailers

The original logic parsed the message in terms of "sections" starting
with a pattern like "<key>: ....", and completely discarded sections
whose keys were defined in the 'MessageSection' enum.

This caused all kinds of problems like:

- subject lines starting with "<something>: ..." were discarded

- entire blocks of text in the body were discarded if there was
  a pattern like "<something>: ...". For example, a line starting
  with "https://..." was discarded.

- all commit trailers with keys not in the 'MessageSection' enum
  were discarded (e.g. "Fixes: ...") were silently discarded;

Behavior Change
---------------

parse_message() now relies on commit_message::parse_commit_message(),
uses the 'git interpret-trailers --parse' command to parse the
trailers section.

parse_commit_message() parses the message in 3 parts: subject,
body and trailers, where the body exclude both the subect and
the trailers.

The new logic has the following implications that breaks with
the previous behavior:

- The **entire** body (i.e. excluding just the subject and all trailers)
  is now stored as the 'MessageSection::Summary' section.

  Notice: we no longer require the "Summary:" section header (which
  was strangely discarded when spr re-wrote the message anyay.)

- All other known sections (eg., "Test-Plan:", "Reviewers:", etc.) now
  **MUST** adhere to the official trailers syntax, which means:

  - They must be grouped in the last paragraph of the message. I.e.,
    if they're not in the last paragraph, they're treated as part
    of the summary section.

  - In multi-line trailers, the subsequent lines must be indented
    by at least one space;

- Trailers unknown to spr are no longer discarded.

Implementation
--------------

We keep track of the unknown trailers using a "pseudo-section"
called 'ExtraTrailers'.

This "pseudo-section" approach is a little hacky. Ideally, we should
modify parse_message()'s return type to a struct with a separate
field just for the extra trailers.

There are however lots of places that use the returned type
(MessageSectionsMap), so modifying it would require some pretty
substantial refactoring.

This change, while not ideal, avoids this larger refactoring and
is completely transparent to the rest of the code.

This `ExtraTrailers` "section" will be used by the build_message()
function in a subsequent change.

We'll modify the build_message() function in a subsequent change to
render the trailers properly.

Pull-Request: https://github.com/aristanetworks/cordspr/pull/10