arista: make build_message() compliant with git trailers syntax
The original rendering of the build message was not compliant with
the git commit trailers syntax.

Now all trailers (known and unknown sections) are grouped in the
last paragraph of the message. Known trailers are added first in
the requested order, and the unkown trailers are added after that
preserving their original order. Also, in multi-line trailers, the
subsequent lines are indented by one space.

Pull-Request: https://github.com/aristanetworks/cordspr/pull/11